Move the pypim test in its own module to be run by the CI
